requirements.txt
argon2-cffi==20.1.0
click==7.1.2
croniter==0.3.36
cryptography==3.3.2
diff-match-patch==20200713
django-anymail==7.2.1
django-compressor==2.4
django-contrib-comments==1.9.2
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.4.0
django-markdownx==3.0.1
django-recaptcha==2.0.6
django-redis==4.11.0
django-reversion-compare==0.12.2
django-reversion==3.0.8
django-rq==2.3.0
django-select2==7.4.2
django-webtest==1.9.7
Django==2.2.20
djangorestframework-csv==2.1.0
djangorestframework==3.12.2
drf-nested-routers==0.92.1
Faker==4.14.2
fakeredis==1.4.4
lupa==1.9
Markdown==3.3.3
mkdocs==1.1.2
Pillow==8.1.1
psycopg2==2.8.6
PyGithub==1.53
PyYAML==5.4
redis==3.4.1
requests-mock==1.8.0
requests==2.25.0
rq-scheduler==0.9.1
rq==1.2.2
social-auth-app-django==4.0.0
social-auth-core==3.3.3
django-formset-js-improved==0.5.0.2
